requirements.txt
django-haystack>=2,<2.6
elasticsearch<5